# api/openapi.yaml
openapi: 3.1.0
info:
  title: llm-proxy web API
  version: 0.1.0
  description: |
    The browser-facing API of llm-proxy: the user cabinet and the admin panel.
    This is NOT the proxied LLM API (`/v1/...`), which the embedded upstream serves
    and which authenticates with API tokens, not sessions.

    Conventions:
    - Authentication is a session cookie set by `POST /api/auth/login` or the OIDC
      callback. It is `HttpOnly`, so no endpoint ever returns a session id.
    - A restricted session (temporary password not yet changed) may call only
      `GET /api/me`, `POST /api/auth/password` and `POST /api/auth/logout`; every
      other authenticated endpoint answers 403 with code `password_change_required`.
    - An authenticated endpoint called without a valid session answers 401 with code
      `unauthenticated`. The only other 401 code is `invalid_credentials`, and it is
      returned only by `POST /api/auth/login` and `POST /api/auth/password`, where it
      rejects what was typed and leaves the session alone: a client must not treat it
      as a signed-out session.
    - Every error the web API writes, including those from its middleware (session,
      restriction, content type), is a JSON `Error` - never plain text.
    - Codes any endpoint may answer besides those it declares: `unsupported_media_type`
      (415, a mutating request without `Content-Type: application/json`),
...
      expects, with `field` when one field is at fault).
...
servers:
  - url: /
...
paths:
...
  /api/admin/users:
    get:
      tags: [admin-users]
      operationId: listUsers
      responses:
        "200":
          description: Every account, humans and service accounts.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminUser" }
    post:
      tags: [admin-users]
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserRequest" }
      responses:
        "201":
          description: >
            Created. For a human with local sign-in, `temporaryPassword` is set and
            shown only here. A service account and an OIDC-only invitation get none.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreatedUser" }
        "409":
          description: "`email_taken` — compared case-insensitively."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: >
            `invalid_input` with `field` naming the offending field (kind, email, role,
            signIn, displayName), or `invalid_rule` with `field` set to the first bad rule.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/admin/users/{userId}:
    get:
      tags: [admin-users]
      operationId: getUser
      parameters: [{ $ref: "#/components/parameters/UserId" }]
      responses:
        "200":
          description: One account.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
    patch:
      tags: [admin-users]
      operationId: updateUser
      parameters: [{ $ref: "#/components/parameters/UserId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateUserRequest" }
      responses:
        "200":
          description: Updated. A policy change applies on the user's next request.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
        "409":
          description: >
            `policy_managed_by_idp` — the policy is recomputed at every IdP login and cannot be
            edited here while the group mapping is configured. `self_lockout` — an administrator
            tried to block themselves or remove their own admin role.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: "`invalid_rule` with `field` set to the offending rule."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/admin/users/{userId}/password-reset:
    post:
      tags: [admin-users]
      operationId: resetUserPassword
      summary: Issue a new temporary password. The user must change it at next sign-in.
      parameters: [{ $ref: "#/components/parameters/UserId" }]
      responses:
        "200":
          description: The temporary password, shown only here.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TemporaryPassword" }
        "409":
          description: "`not_local` — a service account has no password."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/admin/users/{userId}/invitation:
    post:
      tags: [admin-users]
      operationId: renewUserInvitation
      summary: >
        Issue a fresh identity-provider invitation for a human who has not linked an
        identity yet (the old one lapsed or was consumed without a link). Replaces any
        previous invitation for that account.
      parameters: [{ $ref: "#/components/parameters/UserId" }]
      responses:
        "204": { description: Invited. The first IdP sign-in with the account's address claims it. }
        "409":
          description: >
            `already_linked` — the account already has an identity-provider link;
            `not_invitable` — a service account, an account without an email address, or
            identity-provider sign-in is not configured.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/admin/users/{userId}/tokens:
    get:
      tags: [admin-users]
      operationId: listUserTokens
      parameters: [{ $ref: "#/components/parameters/UserId" }]
      responses:
        "200":
          description: That account's tokens.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Token" }
    post:
      tags: [admin-users]
      operationId: issueUserToken
      summary: Issue a token on behalf of an account — the only way a service account gets one.
      parameters: [{ $ref: "#/components/parameters/UserId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IssueTokenRequest" }
      responses:
        "201":
          description: Issued. `secret` is shown here and never again.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IssuedToken" }
        "422":
          description: "`invalid_input` with `field: label` — empty, longer than 64 characters, or with control characters."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: "`token_limit` — the account already holds 50 tokens that are not revoked; revoke one first."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/admin/users/{userId}/tokens/{tokenId}:
    delete:
      tags: [admin-users]
      operationId: revokeUserToken
      parameters:
        - $ref: "#/components/parameters/UserId"
        - $ref: "#/components/parameters/TokenId"
      responses:
        "204": { description: Revoked. }
        "404":
          description: "`not_found` — no such token on that account."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/admin/users/{userId}/activity:
    get:
      tags: [admin-users]
      operationId: getUserActivity
      parameters:
        - $ref: "#/components/parameters/UserId"
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        "200":
          description: Recent proxied requests and recent audit events.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Activity" }
...
  /api/admin/policy/preview:
    post:
      tags: [admin-users]
      operationId: previewPolicy
      summary: >
        Which models of today's catalogue a set of rules covers. A wildcard rule also
        covers models published later; the preview can only show today's.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PolicyPreviewRequest" }
      responses:
        "200":
          description: Coverage and per-rule parse errors.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PolicyPreview" }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: llmproxy_session
  parameters:
    UserId: { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
    TokenId: { name: tokenId, in: path, required: true, schema: { type: string, format: uuid } }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine code. The client translates this, not `message`.
          examples: [invalid_credentials]
        message: { type: string, description: English diagnostic for logs. Not for display. }
        field: { type: string, description: "The request field the error is about, when there is one." }
...
    Kind: { type: string, enum: [human, service] }
    Role: { type: string, enum: [user, admin] }
    Status: { type: string, enum: [active, blocked] }
    PolicySource:
      type: string
      enum: [local, idp]
      description: "`idp` means the policy is recomputed from group claims at every sign-in and is read-only here."
    Policy:
      type: array
      description: >
        Allow-list rules `<provider>:<model-glob>`. The glob dialect is exactly `*`
        (any run of characters, including `/` and `:`) and `?` (one character);
        everything else is literal. A wildcard covers models published later.
      items: { type: string, examples: ["chatgpt:*", "claude:claude-sonnet-5"] }
...
    Token:
      type: object
      required: [id, label, prefix, createdAt]
      properties:
        id: { type: string, format: uuid }
        label: { type: string }
        prefix: { type: string, description: "The first characters of the secret, to tell keys apart." }
        createdAt: { type: string, format: date-time }
        lastUsedAt:
          type: [string, "null"]
          format: date-time
          description: Null if the token was never used.
        revokedAt: { type: [string, "null"], format: date-time }
    IssueTokenRequest:
      type: object
      required: [label]
      properties:
        label: { type: string, minLength: 1, maxLength: 64, description: "Printable text; control characters are refused." }
    IssuedToken:
      type: object
      required: [token, secret]
      properties:
        token: { $ref: "#/components/schemas/Token" }
        secret: { type: string, description: The full API token. Shown once. }
...
    AdminUser:
      type: object
      required: [id, kind, displayName, role, status, policy, policySource, mustChangePassword, signIn, createdAt]
      properties:
        id: { type: string, format: uuid }
        kind: { $ref: "#/components/schemas/Kind" }
        email: { type: [string, "null"] }
        displayName: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        status: { $ref: "#/components/schemas/Status" }
        policy: { $ref: "#/components/schemas/Policy" }
        policySource: { $ref: "#/components/schemas/PolicySource" }
        mustChangePassword: { type: boolean }
        signIn:
          type: array
          description: >
            Sign-in methods that work now: `password` when a usable (unexpired) password is
            set, `oidc` when an identity-provider identity is linked. Empty for a service
            account and for a person whose only way in is a pending invitation.
          items: { type: string, enum: [password, oidc] }
        invitationExpiresAt:
          type: [string, "null"]
          format: date-time
          description: >
            Set while an identity-provider invitation exists and no identity is linked yet;
            it may lie in the past (lapsed - renew it with POST .../invitation).
        lastSeenAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    CreateUserRequest:
      type: object
      required: [kind, displayName, policy]
      properties:
        kind: { $ref: "#/components/schemas/Kind" }
        email: { type: string, description: Required for a human. Absent for a service account. }
        displayName: { type: string, minLength: 1 }
        role: { $ref: "#/components/schemas/Role" }
        policy: { $ref: "#/components/schemas/Policy" }
        signIn:
          type: string
          enum: [password, oidc]
          description: >
            For a human only. `password` issues a temporary password; `oidc` records an
            invitation the first sign-in with that address through the identity provider
            will claim. Ignored for a service account.
    CreatedUser:
      type: object
      required: [user]
      properties:
        user: { $ref: "#/components/schemas/AdminUser" }
        temporaryPassword: { $ref: "#/components/schemas/TemporaryPassword" }
    TemporaryPassword:
      type: object
      required: [password, expiresAt]
      properties:
        password: { type: string, description: Shown once. }
        expiresAt: { type: string, format: date-time }
    UpdateUserRequest:
      type: object
      properties:
        displayName: { type: string, minLength: 1 }
        role: { $ref: "#/components/schemas/Role" }
        status: { $ref: "#/components/schemas/Status" }
        policy: { $ref: "#/components/schemas/Policy" }
    Activity:
      type: object
      required: [requests, audit]
      properties:
        requests:
          type: array
          items:
            type: object
            required: [at, provider, model, stream, statusCode, tokensTotal, latencyMs]
            properties:
              at: { type: string, format: date-time }
              tokenId: { type: [string, "null"], format: uuid }
              provider: { type: string }
              model: { type: string }
              stream: { type: boolean }
              statusCode: { type: integer }
              tokensTotal: { type: integer }
              latencyMs: { type: integer }
        audit:
          type: array
          items:
            type: object
            required: [at, action]
            properties:
              at: { type: string, format: date-time }
              action: { type: string, examples: [token.issue, token.revoke, user.policy.update, auth.sign_in] }
              target: { type: string }
              actorId: { type: [string, "null"], format: uuid }
              detail: { type: object, additionalProperties: true }
...
    PolicyPreviewRequest:
      type: object
      required: [rules]
      properties:
        rules: { $ref: "#/components/schemas/Policy" }
    PolicyPreview:
      type: object
      required: [errors, covered]
      properties:
        errors:
          type: array
          items:
            type: object
            required: [rule, code]
            properties:
              rule: { type: string }
              code: { type: string, examples: [invalid_rule] }
        covered:
          type: array
          description: Models of today's catalogue the rules allow.
          items:
            type: object
            required: [provider, model]
            properties:
              provider: { type: string }
              model: { type: string }
...
security:
  - session: []